pip_requirements.txt
aiohappyeyeballs==2.4.3
aiohttp==3.11.2
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.6.2.post1
appnope==0.1.4
asttokens==2.4.1
attrs==24.2.0
cachetools==5.5.0
certifi==2024.8.30
charset-normalizer==3.4.0
coloredlogs==15.0.1
comm==0.2.2
dataclasses-json==0.6.7
debugpy==1.8.8
decorator==5.1.1
distro==1.9.0
docstring_parser==0.16
et_xmlfile==2.0.0
executing==2.1.0
frozenlist==1.5.0
google-api-core==2.23.0
google-auth==2.36.0
google-cloud-aiplatform==1.72.0
google-cloud-bigquery==3.27.0
google-cloud-core==2.4.1
google-cloud-resource-manager==1.13.1
google-cloud-storage==2.18.2
google-crc32c==1.6.0
google-resumable-media==2.7.2
googleapis-common-protos==1.66.0
grpc-google-iam-v1==0.13.1
grpcio==1.67.1
grpcio-status==1.67.1
h11==0.14.0
httpcore==1.0.6
httpx==0.27.2
httpx-sse==0.4.0
humanfriendly==10.0
idna==3.10
ipykernel==6.29.5
ipython==8.29.0
jedi==0.19.2
jiter==0.7.1
joblib==1.4.2
jsonpatch==1.33
jsonpointer==3.0.0
jupyter_client==8.6.3
jupyter_core==5.7.2
langchain==0.3.7
langchain-community==0.3.7
langchain-core==0.3.18
langchain-google-vertexai==2.0.7
langchain-openai==0.2.8
langchain-redis==0.1.1
langchain-text-splitters==0.3.2
langsmith==0.1.143
marshmallow==3.23.1
matplotlib-inline==0.1.7
ml-dtypes==0.4.1
multidict==6.1.0
mypy-extensions==1.0.0
nest-asyncio==1.6.0
numpy==1.26.4
openai==1.54.4
openpyxl==3.1.5
orjson==3.10.11
packaging==24.2
pandas==2.2.3
parso==0.8.4
pexpect==4.9.0
platformdirs==4.3.6
prompt_toolkit==3.0.48
propcache==0.2.0
proto-plus==1.25.0
protobuf==5.28.3
psutil==6.1.0
ptyprocess==0.7.0
pure_eval==0.2.3
pyasn1==0.6.1
pyasn1_modules==0.4.1
pydantic==2.9.2
pydantic-settings==2.6.1
pydantic_core==2.23.4
Pygments==2.18.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-ulid==2.7.0
pytz==2024.2
PyYAML==6.0.2
pyzmq==26.2.0
redis==5.2.0
redisvl==0.3.6
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
rsa==4.9
scikit-learn==1.5.2
scipy==1.14.1
shapely==2.0.6
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.35
stack-data==0.6.3
tabulate==0.9.0
tenacity==9.0.0
threadpoolctl==3.5.0
tiktoken==0.8.0
tornado==6.4.1
tqdm==4.67.0
traitlets==5.14.3
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
wcwidth==0.2.13
yarl==1.17.1
